+++ packages/frontend-plugin-api/src/index.ts
@@ -20,6 +20,7 @@
  * @packageDocumentation
  */
 
+export * from './apis';
 export * from './components';
 export * from './extensions';
 export * from './routing';