+++ plugins/search-backend-module-elasticsearch/config.d.ts
@@ -115,6 +115,19 @@ export interface Config {
              * Eg. https://my-es-cluster.eu-west-1.es.amazonaws.com
              */
             node: string;
+
+            /**
+             * The AWS region.
+             * Only needed if using a custom DNS record.
+             */
+            region?: string;
+
+            /**
+             * The AWS service used for request signature.
+             * Either 'es' for "Managed Clusters" or 'aoss' for "Serverless".
+             * Only needed if using a custom DNS record.
+             */
+            service?: 'es' | 'aoss';
           }
 
         /**