+++ packages/core-compat-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/core-compat-api
 
+## 0.0.1-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.3.0-next.2
+
 ## 0.0.1-next.1
 
 ### Patch Changes