+++ plugins/apache-airflow/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-apache-airflow
 
+## 0.2.17-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.8-next.2
+
 ## 0.2.17-next.1
 
 ### Patch Changes