+++ plugins/catalog-backend-module-bitbucket/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-catalog-backend-module-bitbucket
 
+## 0.2.22-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.9-next.2
+  - @backstage/plugin-catalog-node@1.5.0-next.2
+
 ## 0.2.22-next.1
 
 ### Patch Changes