+++ plugins/example-todo-list/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @internal/plugin-todo-list
 
+## 1.0.19-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.8-next.2
+
 ## 1.0.19-next.1
 
 ### Patch Changes