+++ plugins/tech-insights-backend-module-jsonfc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-tech-insights-backend-module-jsonfc
 
+## 0.1.39-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.9-next.2
+  - @backstage/plugin-tech-insights-node@0.4.13-next.2
+
 ## 0.1.39-next.1
 
 ### Patch Changes